Member changes
* Sprite imageWidth and imageHeight are now ushorts
* TextField extends DisplayObject instead of Sprite now, has its own
unique pivot settings

+++ Display/Sprite.cs
@@ -4,8 +4,8 @@ namespace Poncho.Display
 {
 	public class Sprite : DisplayObjectContainer
 	{
-		public int imageWidth { get { return image?.rect.width ?? 0; } }
-		public int imageHeight { get { return image?.rect.height ?? 0; } }
+		public ushort imageWidth { get { return image?.rect.width ?? 0; } }
+		public ushort imageHeight { get { return image?.rect.height ?? 0; } }
 		public ushort pivotX { get { return image?.pivot.x ?? 0; } }
 		public ushort pivotY { get { return image?.pivot.y ?? 0; } }
 		public Image image;
